Allow users to customize the opened url
While the hostname of the URL is pulled from the `git ls-remote`
command, the protocol is hardcoded to "https" and the remote used is
hardcoded to nothing (which defaults to origin).  This allows
customizing both of those if the github host (say, an Enterprise
installation) is using a different protocol and if you don't want to
show origin but a different branch.