Use the remote branch name
Without this, the plugin opens github with the local branch name. If
the remote branch name is different, that will result in a 404 on
GitHub.

This also updates the plugin to use the remote associated with the
branch, which won't necessarily be the same as the default remote for
the repo (which was previously used). If a remote is specified in the
user's config, that will still take precedence.